fix: rename generate-post node
